Add option to select webcam driver and fix npe in PipelineSnapshot
