+++ Common/src/main/java/com/github/serivesmejia/eocvsim/util/fps/FpsLimiter.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2021 Sebastian Erives
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ */
+
+package com.github.serivesmejia.eocvsim.util.fps
+
+class FpsLimiter(var maxFPS: Double = 30.0) {
+
+    @Volatile private var start = 0.0
+    @Volatile private var diff = 0.0
+    @Volatile private var wait = 0.0
+
+    @Throws(InterruptedException::class)
+    fun sync() {
+        wait = 1.0 / (maxFPS / 1000.0)
+        diff = System.currentTimeMillis() - start
+        if (diff < wait) {
+            Thread.sleep((wait - diff).toLong())
+        }
+        start = System.currentTimeMillis().toDouble()
+    }
+
+}
