+++ EOCV-Sim/src/main/java/com/github/serivesmejia/eocvsim/gui/util/WebcamDriver.kt
@@ -1,0 +1,5 @@
+package com.github.serivesmejia.eocvsim.gui.util
+
+enum class WebcamDriver {
+    OpenIMAJ, OpenCV
+}
