+++ EOCV-Sim/src/main/java/com/github/serivesmejia/eocvsim/input/camera/openimaj/OpenIMAJWebcam.kt
@@ -93,13 +93,12 @@ class OpenIMAJWebcam @JvmOverloads constructor(
     private var streamThread: Thread? = null
 
     private var frameMat: Mat? = null
-    private val closeLock = Any()
+    private var closeLock = Any()
 
     override fun open() {
         assertNotOpen("open camera")
 
         try {
-
             val capture = VideoCapture(
                 resolution.width.toInt(),
                 resolution.height.toInt(),
@@ -121,6 +120,8 @@ class OpenIMAJWebcam @JvmOverloads constructor(
                 resolution.width.toInt(),
                 CvType.CV_8UC3
             )
+
+            closeLock = Any()
 
             // creating webcam stream and starting it in another thread
             stream = WebcamStream(this, closeLock)
@@ -162,6 +163,7 @@ class OpenIMAJWebcam @JvmOverloads constructor(
         synchronized(closeLock) {
             videoCapture!!.stopCapture()
             videoCapture = null
+            streamThread!!.interrupt()
         }
     }
 
@@ -170,7 +172,6 @@ class OpenIMAJWebcam @JvmOverloads constructor(
         val lock: Any,
         matQueueSize: Int = 2,
     ) : Runnable {
-
         private val grabber = webcam.grabber!!
         private val fpsLimiter = FpsLimiter(webcam.fps)
 
@@ -189,8 +190,28 @@ class OpenIMAJWebcam @JvmOverloads constructor(
 
             while (!Thread.interrupted() && webcam.isOpen && webcam.videoCapture!!.hasNextFrame()) {
                 synchronized(lock) {
+                    // When closing the camera, main thread claims the lock and takes a little while.
+                    // Due to this, the while loop continues to the next iteration and waits for the
+                    // close method to free the lock. Once it's freed and the camera has been closed,
+                    // the iteration continues even after the thread has been interrupted.
+                    //
+                    // This causes grabber.nextFrame to be called below and hang for a little before
+                    // it timeouts and ends the thread, but in some apparently random case, it can
+                    // cause a native jvm crash for some reason.
+                    //
+                    // We check if the thread has been interrupted right after the lock is claimed,
+                    // and if it was, break out of the loop immediately by interrupting the thread
+                    // and exiting from synchronized() to make sure we end the thread right after
+                    // close() is called and the lock is reclaimed. Thanks for coming to my ted talk.
+                    if(Thread.interrupted()) {
+                        Thread.currentThread().interrupt()
+                        return@synchronized // break; because the thread has been interrupted
+                    }
+
                     val err = grabber.nextFrame()
 
+                    // copied from the original source code of VideoCapture.
+                    // no idea what the < -1 error codes stand for.
                     if (err == -1) {
                         Log.warn(TAG, "Timed out waiting for next frame of \"${webcam.name}\"")
                         return@synchronized
@@ -198,12 +219,51 @@ class OpenIMAJWebcam @JvmOverloads constructor(
                         throw RuntimeException("Error occurred getting next frame (code: $err)")
                     }
 
-                    val image = grabber.image ?: return@synchronized
+                    val image = grabber.image ?: return@synchronized // continue; if the image ptr was null
+
+                    val resolution = grabber.resolution
+
+                    // Another instance of an apparently random jvm crash:
+                    // Sometimes the jvm ends unexpectedly with an access violation exit code (0xC0000374)
+                    // i suspect that it's due to the fact that we read more bytes than we have available
+                    // in the call to image.getByteBuffer(), apparently it doesn't implement any sort of
+                    // check so if we overflow it instantly results in a crash.
+                    //
+                    // It *shouldn't* happen since we don't allow the resolution to be changed in the middle of
+                    // a running camera stream, but we implement this check regardless to see if it fixes the issue
+                    if(resolution.width.toInt() != width || resolution.height.toInt() != height) {
+                        Log.warn(TAG, "Grabber currently has a resolution different than the initial one ($resolution vs initial Size($width, $height))")
+                        return@synchronized // continue;
+                    }
+
                     val mat = recycler.takeMat()
 
+                    // most sensitive call in the code: Requesting more bytes than what the native
+                    // Pointer has, will result in a jvm crash. See the check implemented before
                     val buffer = image.getByteBuffer((width * height * 3).toLong())
+
+                    // getByteBuffer returns a Direct ByteBuffer, which means that it directly maps
+                    // to an area of native memory. Calling buffer.get(array) on a direct buffer,
+                    // directly copies the pixels using memcpy under the hood, so this is about the
+                    // fastest we can get with the openimaj driver in pure kotlin. (this could probably
+                    // be reduced to zero or one copies instead of two using native c++ but i don't
+                    // really wanna go through that pain again...)
+                    //
+                    // We reuse the same pixel array every loop so that the gc doesn't panic and lets
+                    // us run smoothly. Benchmarks performed at some point revealed that creating a
+                    // new array every loop caused spikes of up to 300-500ms constantly, reducing our
+                    // frame rate by a non-negligible amount. Thankfully the direct byte buffer created
+                    // every frame doesn't make the gc panic because of the same fact that it's just
+                    // mapped to bytes in the native memory, outside of the gc's managed memory.
                     buffer.get(pixels)
 
+                    // OpenIMAJGrabber already stores pixels under the hood
+                    // in a continuous 1D array in the rgb format, that's
+                    // where (width * height * 3) comes from, the "3" constant
+                    // stands for the three channels in the rgb color space.
+                    //
+                    // That is extremely convenient, it directly allows us to
+                    // copy the pixels array into the opencv mat. Thank goodness.
                     mat.put(0, 0, pixels)
 
                     queue.offer(mat)
