--- EOCV-Sim/src/main/java/com/github/serivesmejia/eocvsim/util/fps/FpsCounter.kt
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2021 Sebastian Erives
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- */
-
-package com.github.serivesmejia.eocvsim.util.fps
-
-import com.qualcomm.robotcore.util.ElapsedTime
-import com.qualcomm.robotcore.util.MovingStatistics
-
-class FpsCounter {
-
-    private val elapsedTime = ElapsedTime()
-
-    private val avgFpsStatistics = MovingStatistics(100)
-
-    val avgFps: Double
-        get() {
-            return avgFpsStatistics.mean
-        }
-
-    @Volatile
-    private var fpsCount = 0
-
-    @get:Synchronized
-    @Volatile
-    var fps = 0
-        private set
-
-    @Synchronized
-    fun update() {
-        fpsCount++
-        if (elapsedTime.seconds() >= 1) {
-            fps = fpsCount; fpsCount = 0
-            avgFpsStatistics.add(fps.toDouble())
-            elapsedTime.reset()
-        }
-    }
-
-}
