+++ EOCV-Sim/src/main/java/org/firstinspires/ftc/robotcore/internal/opmode/TelemetryTransmissionReceiver.kt
@@ -3,5 +3,5 @@ package org.firstinspires.ftc.robotcore.internal.opmode
 import org.firstinspires.ftc.robotcore.external.Telemetry
 
 interface TelemetryTransmissionReceiver {
-    fun onTelemetryTransmission(text: String, currentTelemetry: Telemetry)
+    fun onTelemetryTransmission(text: String, srcTelemetry: Telemetry)
 }
